webhook/requirements.txt
aiogram==3.4.1
pyMongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
geopy==2.4.1
numpy==1.26.4